Extract shared component detail view
